requirements.txt
aws-xray-sdk==0.97
mysql-connector-python==8.0.18
